re-architecture of the application
