--- src/Head.js
@@ -1,5 +1,0 @@
-import React, { Component } from "react";
-
-export default function() {
-  return <h1>Hello world!</h1>;
-}
