+++ src/Header/index.js
@@ -6,8 +6,8 @@ import Nav from "./Nav";
 
 export const Header = styled.header`
   display: flex;
-  height: 80px;
   align-items: center;
+  height: 80px;
   border-bottom: 1px solid rgba(72, 72, 72, 0.3);
   box-shadow: 0 0.5px 0 rgba(72, 72, 72, 0.3);
 `;
