+++ src/Main/common/Image.js
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Image = styled.img`
+  width: 100%;
+  max-height: 346px;
+  object-fit: cover;
+`;
+
+export default function() {
+  return <Image />;
+}
