+++ src/Main/common/Name.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Name = styled.span`
+  font-weight: 600;
+`;
+
+export default function() {
+  return <Name />;
+}
