+++ src/Main/common/Price.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Price = styled.span`
+  font-weight: bold;
+`;
+
+export default function() {
+  return <Price />;
+}
