+++ src/Main/common/Rating.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Rating = styled.div`
+  margin-top: 8px;
+`;
+
+export default function() {
+  return <Rating />;
+}
