+++ src/Main/common/Reviews.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Reviews = styled.span`
+  font-weight: 600;
+  font-size: 12px;
+  line-height: 14px;
+  color: #383838;
+`;
+
+export default function() {
+  return <Reviews />;
+}
