+++ src/Main/common/Slider.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Slider = styled.div`
+  position: relative;
+`;
+
+export default function() {
+  return <Slider />;
+}
