+++ src/Main/common/Star.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Star = styled.img`
+  margin-right: 4px;
+`;
+
+export default function() {
+  return <Star />;
+}
