+++ src/Main/common/Stars.js
@@ -1,0 +1,11 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Stars = styled.div`
+  display: inline-block;
+  margin-right: 8px;
+`;
+
+export default function() {
+  return <Stars />;
+}
