+++ src/Main/common/Title.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+import styled from "styled-components";
+
+export const Title = styled.div`
+  margin-top: 8px;
+  font-size: 15px;
+  line-height: 18px;
+  color: #383838;
+`;
+
+export default function() {
+  return <Title />;
+}
